requirements.txt
faiss-gpu-cu12
sentence-transformers
torch
tqdm
numpy
langchain
langchain_community
pandas
accelerate
datasets
mistral_inference
bitsandbytes
langchain-core
openai
matplotlib
seaborn
google-genai
langchain_community